requirements_pip.txt
torch==2.4.1 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
accelerate==1.1.1
bertopic
optuna
bitsandbytes==0.44.1
triton==3.0.0
peft==0.17.1
sentence_transformers